# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Billing Service API",
    "version": "0.1.0",
    "description": "Manages billing accounts, credit balance, deductions, and Stripe integration for the Conductor platform."
  },
  "servers": [
    {
      "url": "http://localhost:3012"
    }
  ],
  "components": {
    "schemas": {
...
      "BillingAccount": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "org_id": {
            "type": "string",
            "format": "uuid"
          },
          "credited_cents": {
            "type": "string"
          },
          "usage_cents": {
            "type": "string",
            "description": "Platform usage from runs-service, including actualized costs and provisioned holds. Use with balance_cents for spend authorization."
          },
          "balance_cents": {
            "type": "string",
            "description": "Spendable funds: credited_cents minus usage_cents. Includes provisioned holds, so this is the safety value for authorization, depletion, runway, and top-up checks."
          },
          "actual_balance_cents": {
            "type": "string",
            "description": "User-facing credit balance: credited funds minus actualized platform usage only. Provisioned holds are not subtracted here because they may later actualize or cancel."
          },
          "usage_discount_pct": {
            "type": "integer",
            "nullable": true
          },
          "net_usage_cents": {
            "type": "string"
          },
          "net_actual_usage_cents": {
            "type": "string"
          },
          "topup_amount_cents": {
            "type": "integer",
            "nullable": true
          },
          "topup_threshold_cents": {
            "type": "integer",
            "nullable": true
          },
          "has_payment_method": {
            "type": "boolean"
          },
          "has_auto_topup": {
            "type": "boolean"
          },
          "auto_reload_supported": {
            "type": "boolean"
          },
          "auto_reload_unsupported_reason": {
            "type": "string",
            "nullable": true
          },
          "card_country": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "org_id",
          "credited_cents",
          "usage_cents",
          "balance_cents",
          "actual_balance_cents",
          "usage_discount_pct",
          "net_usage_cents",
          "net_actual_usage_cents",
          "topup_amount_cents",
          "topup_threshold_cents",
          "has_payment_method",
          "has_auto_topup",
          "auto_reload_supported",
          "auto_reload_unsupported_reason",
          "card_country",
          "created_at",
          "updated_at"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "BalanceResponse": {
        "type": "object",
        "properties": {
          "balance_cents": {
            "type": "string",
            "description": "Spendable funds: credited_cents minus usage_cents. Includes provisioned holds, so this is the safety value for authorization, depletion, runway, and top-up checks."
          },
          "actual_balance_cents": {
            "type": "string",
            "description": "User-facing credit balance: credited funds minus actualized platform usage only. Provisioned holds are not subtracted here because they may later actualize or cancel."
          },
          "depleted": {
            "type": "boolean"
          }
        },
        "required": [
          "balance_cents",
          "actual_balance_cents",
          "depleted"
        ]
      },
      "UpdateAutoTopupRequest": {
        "type": "object",
        "properties": {
          "topup_amount_cents": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true
          },
          "topup_threshold_cents": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "topup_amount_cents",
          "topup_threshold_cents"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/v1/accounts": {
      "get": {
        "summary": "Get or create billing account for org",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-api-key",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "x-run-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Campaign ID injected by workflow-service"
            },
            "required": false,
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Brand ID(s) injected by workflow-service (comma-separated UUIDs for multi-brand campaigns)",
              "example": "uuid1,uuid2,uuid3"
            },
            "required": false,
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow slug injected by workflow-service"
            },
            "required": false,
            "name": "x-workflow-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Feature slug for tracking"
            },
            "required": false,
            "name": "x-feature-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Audience ID injected by campaign-service for per-audience cost attribution"
            },
            "required": false,
            "name": "x-audience-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Billing account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingAccount"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "stripe-service or runs-service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/accounts/balance": {
      "get": {
        "summary": "Quick balance check (spendable funds)",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-api-key",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "x-run-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Campaign ID injected by workflow-service"
            },
            "required": false,
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Brand ID(s) injected by workflow-service (comma-separated UUIDs for multi-brand campaigns)",
              "example": "uuid1,uuid2,uuid3"
            },
            "required": false,
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow slug injected by workflow-service"
            },
            "required": false,
            "name": "x-workflow-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Feature slug for tracking"
            },
            "required": false,
            "name": "x-feature-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Audience ID injected by campaign-service for per-audience cost attribution"
            },
            "required": false,
            "name": "x-audience-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Balance info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalanceResponse"
                }
              }
            }
          },
          "404": {
            "description": "Billing account not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/accounts/auto_topup": {
      "patch": {
        "summary": "Configure auto-topup settings (requires payment method via stripe-service)",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-api-key",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "x-run-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Campaign ID injected by workflow-service"
            },
            "required": false,
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Brand ID(s) injected by workflow-service (comma-separated UUIDs for multi-brand campaigns)",
              "example": "uuid1,uuid2,uuid3"
            },
            "required": false,
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow slug injected by workflow-service"
            },
            "required": false,
            "name": "x-workflow-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Feature slug for tracking"
            },
            "required": false,
            "name": "x-feature-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Audience ID injected by campaign-service for per-audience cost attribution"
            },
            "required": false,
            "name": "x-audience-id",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAutoTopupRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingAccount"
                }
              }
            }
          },
          "400": {
            "description": "Payment method required or invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Billing account not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Disable auto-topup",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-api-key",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "x-run-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Campaign ID injected by workflow-service"
            },
            "required": false,
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Brand ID(s) injected by workflow-service (comma-separated UUIDs for multi-brand campaigns)",
              "example": "uuid1,uuid2,uuid3"
            },
            "required": false,
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow slug injected by workflow-service"
            },
            "required": false,
            "name": "x-workflow-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Feature slug for tracking"
            },
            "required": false,
            "name": "x-feature-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Audience ID injected by campaign-service for per-audience cost attribution"
            },
            "required": false,
            "name": "x-audience-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Updated account with auto-topup disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingAccount"
                }
              }
            }
          },
          "404": {
            "description": "Billing account not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}