# src/routes/accounts.ts
import { Router } from "express";
import { eq } from "drizzle-orm";
import crypto from "crypto";
import { db } from "../db/index.js";
import { billingAccounts } from "../db/schema.js";
import { requireOrgHeaders, getWorkflowHeaders, forwardWorkflowHeaders } from "../middleware/auth.js";
import { UpdateAutoTopupRequestSchema, WalletSetupRequestSchema } from "../schemas.js";
import { findOrCreateAccount, findOrCreateWalletAccount } from "../lib/account.js";
import { addCents, isDepleted, subCents } from "../lib/cents.js";
import { tierFor } from "../lib/topup-tier.js";
import { fetchRunsOrgActualUsageTotal, fetchRunsOrgUsageTotal } from "../lib/runs-client.js";
import { grantFirstLoadMatch, sumLocalPromoCreditsForOrg } from "../lib/promos.js";
import { getUsageDiscountPct, applyUsageDiscount } from "../lib/usage-discount.js";
import { reloadViaPaymentIntent } from "../lib/reload.js";
import {
  getCustomerByOrg,
  sumSucceededTopupsForCustomer,
  hasAttachedCardPm,
  getOrgCardCountry,
  isAutoReloadBlockedCountry,
} from "../lib/stripe-service-client.js";

const router = Router();
const INITIAL_LOAD_TIMEOUT_MS = 30_000;
const INITIAL_LOAD_IDEMPOTENCY_BUCKET_MS = 60_000;

function buildIdentity(
  orgId: string,
  userId: string,
  runId: string | undefined,
  wfHeaders: Record<string, string>
): Record<string, string> {
  const out: Record<string, string> = {
    "x-org-id": orgId,
    "x-user-id": userId,
    ...wfHeaders,
  };
  if (runId) out["x-run-id"] = runId;
  return out;
}

async function composeAccountFunds(
  orgId: string,
  identity: Record<string, string>
): Promise<{
  creditedCents: string;
  usageCents: string;
  balanceCents: string;
  actualBalanceCents: string;
  discountPct: number | null;
  netUsageCents: string;
  netActualUsageCents: string;
  paidTopupsCents: string;
  hasPaymentMethod: boolean;
  cardCountry: string | null;
  autoReloadSupported: boolean;
}> {
  const customer = await getCustomerByOrg(identity);
  const [paidTopups, localCredits, runsUsage, actualRunsUsage, hasCardPm, cardCountry, discountPct] =
    await Promise.all([
      sumSucceededTopupsForCustomer(identity, customer.id),
      sumLocalPromoCreditsForOrg(orgId),
      fetchRunsOrgUsageTotal(orgId, identity),
      fetchRunsOrgActualUsageTotal(orgId, identity),
      hasAttachedCardPm(identity, customer.id),
      getOrgCardCountry(identity, customer.id),
      getUsageDiscountPct(orgId),
    ]);
  const creditedCents = addCents(paidTopups, localCredits);
  // Discount reduces the usage subtracted from credited (gross usage_cents stays
  // reporting truth). Both net values equal their gross when discountPct is null,
  // so balance_cents / actual_balance_cents are byte-identical without a discount.
  const netUsageCents = applyUsageDiscount(runsUsage.spent_cents, discountPct);
  const netActualUsageCents = applyUsageDiscount(actualRunsUsage.spent_cents, discountPct);
  const balanceCents = subCents(creditedCents, netUsageCents);
  const actualBalanceCents = subCents(creditedCents, netActualUsageCents);
  return {
    creditedCents,
    usageCents: runsUsage.spent_cents,
    balanceCents,
    actualBalanceCents,
    discountPct,
    netUsageCents,
    netActualUsageCents,
    paidTopupsCents: paidTopups,
    hasPaymentMethod: hasCardPm,
    cardCountry,
    autoReloadSupported: !isAutoReloadBlockedCountry(cardCountry),
  };
}

function buildAccountResponse(
  account: typeof billingAccounts.$inferSelect,
  funds: {
    creditedCents: string;
    usageCents: string;
    balanceCents: string;
    actualBalanceCents: string;
    discountPct: number | null;
    netUsageCents: string;
    netActualUsageCents: string;
    paidTopupsCents: string;
    hasPaymentMethod: boolean;
    cardCountry: string | null;
    autoReloadSupported: boolean;
  }
) {
  // Auto-topup is enabled iff the stored columns are non-null (they are the
  // enabled flag). When enabled, the effective (amount, threshold) are the
  // DERIVED postpaid tier (a function of cumulative paid topups) — the negative
  // threshold is the credit-line floor the dashboard renders. When disabled,
  // both fields are null (no top-up). See lib/topup-tier.
  const enabled =
    account.topupAmountCents != null && account.topupThresholdCents != null;
  const tier = enabled ? tierFor(funds.paidTopupsCents) : null;
  return {
    id: account.id,
    org_id: account.orgId,
    credited_cents: funds.creditedCents,
    usage_cents: funds.usageCents,
    balance_cents: funds.balanceCents,
    actual_balance_cents: funds.actualBalanceCents,
    // Per-org platform-usage discount. null = no discount (usage_cents is the net
    // usage, balance_cents/actual_balance_cents unchanged). When set, usage_cents
    // stays GROSS (reporting) and net_usage_cents / net_actual_usage_cents are the
    // discounted usage the balance figures already subtract — the dashboard renders
    // the banner (usage_discount_pct) + strikethrough (gross usage_cents vs net).
    usage_discount_pct: funds.discountPct,
    net_usage_cents: funds.netUsageCents,
    net_actual_usage_cents: funds.netActualUsageCents,
    topup_amount_cents: tier ? tier.amountCents : null,
    topup_threshold_cents: tier ? tier.thresholdCents : null,
    has_payment_method: funds.hasPaymentMethod,
    // Off_session auto-reload is impossible for cards issued in mandate-required countries
    // (e.g. India / RBI, issue #220). When unsupported, the dashboard shows a notice and
    // has_auto_topup is false even if topup config exists — the reload would never fire.
    auto_reload_supported: funds.autoReloadSupported,
    auto_reload_unsupported_reason: funds.autoReloadSupported ? null : "card_issuing_country_unsupported",
    card_country: funds.cardCountry,
    has_auto_topup:
      account.topupAmountCents != null &&
      account.topupThresholdCents != null &&
      funds.hasPaymentMethod &&
      funds.autoReloadSupported,
    created_at: account.createdAt.toISOString(),
    updated_at: account.updatedAt.toISOString(),
  };
}

function initialLoadIdempotencyKey(orgId: string, amountCents: number): string {
  const bucket = Math.floor(Date.now() / INITIAL_LOAD_IDEMPOTENCY_BUCKET_MS);
  return crypto
    .createHash("sha256")
    .update(`wallet_setup:${orgId}:${amountCents}:${bucket}`)
    .digest("hex")
    .slice(0, 32);
}

function withTimeout<T>(ms: number, p: Promise<T>): Promise<T> {
  return new Promise((resolve, reject) => {
    const timer = setTimeout(() => reject(new Error(`initial load timeout after ${ms}ms`)), ms);
    p.then(
      (v) => { clearTimeout(timer); resolve(v); },
      (e) => { clearTimeout(timer); reject(e); }
    );
  });
}

router.get("/v1/accounts", requireOrgHeaders, async (req, res) => {
  try {
    const orgId = req.headers["x-org-id"] as string;
    const userId = req.headers["x-user-id"] as string;
    const runId = req.headers["x-run-id"] as string;
    const wfHeaders = forwardWorkflowHeaders(getWorkflowHeaders(req));
    const identity = buildIdentity(orgId, userId, runId, wfHeaders);

    const account = await findOrCreateAccount(orgId, userId, wfHeaders);

    let funds;
    try {
      funds = await composeAccountFunds(orgId, identity);
    } catch (err) {
      console.error("[billing-service] Failed to compose account funds:", err);
      res.status(502).json({ error: "Failed to compose account funds" });
      return;
    }

    res.json(buildAccountResponse(account, funds));
  } catch (err) {
    console.error("[billing-service] Error getting/creating account:", err);
    res.status(500).json({ error: "Internal server error" });
  }
});

router.get("/v1/accounts/balance", requireOrgHeaders, async (req, res) => {
  try {
    const orgId = req.headers["x-org-id"] as string;
    const userId = req.headers["x-user-id"] as string;
    const runId = req.headers["x-run-id"] as string;
    const wfHeaders = forwardWorkflowHeaders(getWorkflowHeaders(req));
    const identity = buildIdentity(orgId, userId, runId, wfHeaders);

    const [account] = await db
      .select()
      .from(billingAccounts)
      .where(eq(billingAccounts.orgId, orgId))
      .limit(1);

    if (!account) {
      res.status(404).json({ error: "Billing account not found" });
      return;
    }

    let funds;
    try {
      funds = await composeAccountFunds(orgId, identity);
    } catch (err) {
      console.error("[billing-service] Failed to compose account funds:", err);
      res.status(502).json({ error: "Failed to compose account funds" });
      return;
    }

    res.json({
      balance_cents: funds.balanceCents,
      actual_balance_cents: funds.actualBalanceCents,
      depleted: isDepleted(funds.balanceCents),
    });
  } catch (err) {
    console.error("[billing-service] Error checking balance:", err);
    res.status(500).json({ error: "Internal server error" });
  }
});

router.patch("/v1/accounts/auto_topup", requireOrgHeaders, async (req, res) => {
  try {
    const orgId = req.headers["x-org-id"] as string;
    const userId = req.headers["x-user-id"] as string;
    const runId = req.headers["x-run-id"] as string;
    const wfHeaders = forwardWorkflowHeaders(getWorkflowHeaders(req));
    const identity = buildIdentity(orgId, userId, runId, wfHeaders);

    const parsed = UpdateAutoTopupRequestSchema.safeParse(req.body);
    if (!parsed.success) {
      res.status(400).json({ error: parsed.error.message });
      return;
    }
    const { topup_amount_cents, topup_threshold_cents } = parsed.data;

    const [account] = await db
      .select()
      .from(billingAccounts)
      .where(eq(billingAccounts.orgId, orgId))
      .limit(1);

    if (!account) {
      res.status(404).json({ error: "Billing account not found" });
      return;
    }

    let hasCardPm: boolean;
    let cardCountry: string | null;
    try {
      const customer = await getCustomerByOrg(identity);
      [hasCardPm, cardCountry] = await Promise.all([
        hasAttachedCardPm(identity, customer.id),
        getOrgCardCountry(identity, customer.id),
      ]);
    } catch (err) {
      console.error("[billing-service] Failed to fetch customer for PM check:", err);
      res.status(502).json({ error: "Failed to query payment method status" });
      return;
    }

    if (!hasCardPm) {
      res.status(400).json({
        error: "Payment method required. Create a checkout session first.",
      });
      return;
    }

    // Off_session auto-reload can't be charged for cards issued in mandate-required
    // countries (e.g. India / RBI, issue #220). Reject the config rather than store one
    // that silently never fires — fail loud so the dashboard surfaces the real reason.
    if (isAutoReloadBlockedCountry(cardCountry)) {
      res.status(400).json({
        error: `Auto-reload is unavailable for cards issued in ${cardCountry} — off-session charges require a mandate Stripe can't register on this card. Add a card from another country to enable auto-reload.`,
      });
      return;
    }

    const [updated] = await db
      .update(billingAccounts)
      .set({
        topupAmountCents: topup_amount_cents,
        topupThresholdCents: topup_threshold_cents,
        updatedAt: new Date(),
      })
      .where(eq(billingAccounts.orgId, orgId))
      .returning();

    let funds;
    try {
      funds = await composeAccountFunds(orgId, identity);
    } catch (err) {
      console.error("[billing-service] Failed to compose account funds:", err);
      res.status(502).json({ error: "Failed to compose account funds" });
      return;
    }

    res.json(buildAccountResponse(updated, funds));
  } catch (err) {
    console.error("[billing-service] Error updating auto-topup:", err);
    res.status(500).json({ error: "Internal server error" });
  }
});

router.post("/v1/accounts/wallet_setup", requireOrgHeaders, async (req, res) => {
  try {
    const orgId = req.headers["x-org-id"] as string;
    const userId = req.headers["x-user-id"] as string;
    const runId = req.headers["x-run-id"] as string;
    const wfHeaders = forwardWorkflowHeaders(getWorkflowHeaders(req));
    const identity = buildIdentity(orgId, userId, runId, wfHeaders);

    const parsed = WalletSetupRequestSchema.safeParse(req.body);
    if (!parsed.success) {
      res.status(400).json({ error: parsed.error.message });
      return;
    }
    const {
      initial_load_amount_cents,
      topup_amount_cents,
      topup_threshold_cents,
    } = parsed.data;

    await findOrCreateWalletAccount(orgId, userId, wfHeaders);

    let customer;
    let hasCardPm: boolean;
    try {
      customer = await getCustomerByOrg(identity);
      hasCardPm = await hasAttachedCardPm(identity, customer.id);
    } catch (err) {
      console.error("[billing-service] wallet_setup PM check failed:", err);
      res.status(502).json({ error: "Failed to query payment method status" });
      return;
    }

    if (!hasCardPm) {
      res.status(400).json({
        error: "Payment method required. Create a setup checkout session first.",
      });
      return;
    }

    let initialLoad;
    try {
      initialLoad = await withTimeout(
        INITIAL_LOAD_TIMEOUT_MS,
        reloadViaPaymentIntent(
          identity,
          initial_load_amount_cents,
          initialLoadIdempotencyKey(orgId, initial_load_amount_cents),
          { org_id: orgId, billing_reason: "initial_load" }
        )
      );
    } catch (err) {
      console.error("[billing-service] wallet_setup initial load failed:", err);
      res.status(502).json({ error: "Initial load via stripe-service failed" });
      return;
    }

    if (initialLoad.status !== "succeeded") {
      res.status(402).json({
        error: "Initial load payment failed",
        failure_reason: initialLoad.failure_reason ?? "payment_intent_not_succeeded",
      });
      return;
    }

    const [updated] = await db
      .update(billingAccounts)
      .set({
        topupAmountCents: topup_amount_cents,
        topupThresholdCents: topup_threshold_cents,
        updatedAt: new Date(),
      })
      .where(eq(billingAccounts.orgId, orgId))
      .returning();

    const match = await grantFirstLoadMatch(orgId, userId, initial_load_amount_cents);

    let funds;
    try {
      funds = await composeAccountFunds(orgId, identity);
    } catch (err) {
      console.error("[billing-service] Failed to compose account funds:", err);
      res.status(502).json({ error: "Failed to compose account funds" });
      return;
    }

    res.json({
      ...buildAccountResponse(updated, funds),
      initial_load_amount_cents,
      initial_load_payment_intent_id: initialLoad.payment_intent_id,
      first_load_match_applied: match.applied,
      first_load_match_cents: match.amountCents,
      first_load_match_local_promo_id: match.localPromoId,
    });
  } catch (err) {
    console.error("[billing-service] Error setting up wallet:", err);
    res.status(500).json({ error: "Internal server error" });
  }
});

router.delete("/v1/accounts/auto_topup", requireOrgHeaders, async (req, res) => {
  try {
    const orgId = req.headers["x-org-id"] as string;
    const userId = req.headers["x-user-id"] as string;
    const runId = req.headers["x-run-id"] as string;
    const wfHeaders = forwardWorkflowHeaders(getWorkflowHeaders(req));
    const identity = buildIdentity(orgId, userId, runId, wfHeaders);

    const [account] = await db
      .select()
      .from(billingAccounts)
      .where(eq(billingAccounts.orgId, orgId))
      .limit(1);

    if (!account) {
      res.status(404).json({ error: "Billing account not found" });
      return;
    }

    const [updated] = await db
      .update(billingAccounts)
      .set({
        topupAmountCents: null,
        topupThresholdCents: null,
        updatedAt: new Date(),
      })
      .where(eq(billingAccounts.orgId, orgId))
      .returning();

    let funds;
    try {
      funds = await composeAccountFunds(orgId, identity);
    } catch (err) {
      console.error("[billing-service] Failed to compose account funds:", err);
      res.status(502).json({ error: "Failed to compose account funds" });
      return;
    }

    res.json(buildAccountResponse(updated, funds));
  } catch (err) {
    console.error("[billing-service] Error disabling auto-topup:", err);
    res.status(500).json({ error: "Internal server error" });
  }
});

export default router;